refactor(store): split sheetsData logic into rows, columns and cells modules
